Cleaned up positions, implemented physics based jumping

+++ core/src/com/jupiter/rogue/Model/Map/Position.java
@@ -30,7 +30,7 @@ public class Position {
         return yPos;
     }
 
-    public Position(int xPos, int yPos){
+    public Position(float xPos, float yPos){
         this.xPos = xPos;
         this.yPos = yPos;
     }
